refactor: refactor due PR
