+++ src/components/Card/Card.model.ts
@@ -1,0 +1,3 @@
+export type CardProps = {
+  name: string;
+};
