--- src/components/card-wrapper/CardWrapper.scss
@@ -1,3 +1,0 @@
-.CardWrapper{
-
-}
