+++ src/index.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import 'bootstrap/dist/css/bootstrap.min.css';
 import './sass/style.scss';
-import App from './components/app/App';
+import App from './components/App';
 
 ReactDOM.render(
   <React.StrictMode>
